Move mongoDB models to legacy namespace and create delegators
